# System.Diagnostics.Activity Perf Improvement