@CodeBlanch 2020-08-05
Some other areas I noticed:

* `CreateAndStart` method takes in IEnumerable Tags & Links, which it will enumerate. Those will allocate. If we changed the API to accept IList instead of IEnumerable we could do a more simple `for` loop which wouldn't allocate?
* `IEnumerable<KeyValuePair<string, string?>> Tags` property has some special logic to only return the strings in the LinkedList. I couldn't figure out a way to solve that without an allocation. Probably not a big deal, if caller is worried about the perf they can use TagObjects and just exclude the non-string items?
* `IEnumerable<KeyValuePair<string, string?>> Baggage` property is an interesting enumeration. It goes up the Parent chain enumerating over any items it finds along the way. Might be able to craft something for that. For now I left it alone because I don't think in OpenTelemetry we export Baggage meaning there isn't a foreach in the hot path.