@tarekgh 2020-08-05
@CodeBlanch thanks for submitting this. in general LGTM. 
I want to call out one thing which is, this enumeration now will not be thread safe. I think this is fine but we should clarify that in the doc.

> CreateAndStart method takes in IEnumerable Tags & Links, which it will enumerate. Those will allocate. If we changed the API to accept IList instead of IEnumerable we could do a more simple for loop which wouldn't allocate?

I don't think using IList will be a good idea. This will restrict the users of this API to some specific collections and will not be able to use the other collections (e.g. Dictionary<,>)

> IEnumerable<KeyValuePair<string, string?>> Tags property has some special logic to only return the strings in the LinkedList. I couldn't figure out a way to solve that without an allocation. Probably not a big deal, if caller is worried about the perf they can use TagObjects and just exclude the non-string items?

I think this can still be done by having the Enumerator<>.MoveNext to do this logic. no?

> IEnumerable<KeyValuePair<string, string?>> Baggage property is an interesting enumeration. It goes up the Parent chain enumerating over any items it finds along the way. Might be able to craft something for that. For now I left it alone because I don't think in OpenTelemetry we export Baggage meaning there isn't a foreach in the hot path.

agree, we can optimize this later if we need to.