@CodeBlanch 2020-08-05
> I want to call out one thing which is, this enumeration now will not be thread safe. I think this is fine but we should clarify that in the doc.

Can you go into a little more detail on that for me? It looks essentially the same as before to me, so I'm just curious what I'm missing.

> I don't think using IList will be a good idea. This will restrict the users of this API to some specific collections and will not be able to use the other collections (e.g. Dictionary<,>)

Good point. It would be nice to have a solve here, but I can't think of anything. Tags you can apply after creation, so you could workaround it if you were so inclined. But Links, you can only pass on ctor. If we added Add/Remove Link, we could workaround for that too.

> > IEnumerable<KeyValuePair<string, string?>> Tags property has some special logic to only return the strings in the LinkedList. I couldn't figure out a way to solve that without an allocation. Probably not a big deal, if caller is worried about the perf they can use TagObjects and just exclude the non-string items?
> 
> I think this can still be done by having the Enumerator<>.MoveNext to do this logic. no?

It seems like it should be possible, doesn't it? I tried a couple of ways, and just tried again, but I can't get it without an allocation. The method needs to return an IEnumerable which exposes the GetEnumerator. So if I do return `new StringEnumerableTagThing(_tags)` no problem, but there's an allocation. What I really want to do is add `IEnumerable<KeyValuePair<string, string>>` on `TagsLinkedList` but it already has `IEnumerable<KeyValuePair<string, object?>>` so they collide on the `IEnumerable.GetEnumerator` method.