@tarekgh 2020-08-05
> Can you go into a little more detail on that for me? It looks essentially the same as before to me, so I'm just curious what I'm missing.

Sorry I was not clear. you are not changing the thread safety issue. I was just calling out the fact that the enumeration is not thread safe even before your change.

For Tags enumeration, it is ok to keep it like that for now even if it is allocating.

one last ask here, I am seeing the issue https://github.com/dotnet/runtime/issues/40366 which need a little change. can you add this change with your here? at least to avoid the conflicts? thanks.