@tarekgh 2020-08-05
@noahfalk I am merging this but let me know if you have anything you want to change so we can still do it.

@CodeBlanch thanks for your help with this issue.