@noahfalk 2020-08-06
Usually when we make a perf change we also have result from BenchmarkDotNet showing the improvement we got from the change. Can we do that here?

Also just to confirm the new code still allocates, but I think it allocates less than the previous did (old code allocated an IEnumerable and an IEnumerator, new code only allocates the IEnumerator). Does that sound accurate?