@CodeBlanch 2020-08-06
@noahfalk Agreed. For most callers the `struct Enumerator` will be boxed up/allocated because it is accessed through the IEnumerable/IEnumerator interface. The reason I asked for these is we have a ~hack~ [helper](https://github.com/open-telemetry/opentelemetry-dotnet/blob/73bff75ef653f81fe6877299435b21131be36dc0/src/OpenTelemetry/Internal/EnumerationHelper.cs#L96) in OTel which will lookup the `struct GetEnumerator` and build a DynamicMethod for it. So, if you are determined enough, these do actually allow you to get there.

Regarding the perf tests, won't really look as good without the helper. You still want me to do them? Or would you like to see before/after from OTel perspective maybe?