@noahfalk 2020-08-06
> Regarding the perf tests, won't really look as good without the helper. You still want me to do them?

Yes please. I want to make sure that when people look at this change and it claims "we made perf better" then there is some data to substantiate the claim. It doesn't have to look amazing and you are welcome to include extra OTel specific cases that have even larger gains if you want to show that off : )