@CodeBlanch 2020-08-07
@noahfalk Working on the perf tests. It looks like the GetEnumerator I added is being removed. I'm guessing because nothing is referencing it in the code base and it is internal?

If I do this, it seems to work:
```csharp
        [DynamicDependency(DynamicallyAccessedMemberTypes.PublicMethods, typeof(TagsLinkedList))]
        private TagsLinkedList? _tags;
        [DynamicDependency(DynamicallyAccessedMemberTypes.PublicMethods, typeof(LinkedList<ActivityLink>))]
        private LinkedList<ActivityLink>? _links;
        [DynamicDependency(DynamicallyAccessedMemberTypes.PublicMethods, typeof(LinkedList<ActivityEvent>))]
        private LinkedList<ActivityEvent>? _events;
```

Correct approach? OK to PR this change?