@tarekgh 2020-08-07
@eerhardt 

@CodeBlanch has been optimizing the `Activity.TagObject` which return IEnumerable<...>. He provided GetEnumerator there. you can look at this PR. He notice `GetEnumerator` is removed from the code. Is this because of the linker? notice this is in the System.Diagnostics.DiagnosticSource library. 

@CodeBlanch are you using the official build of this library, or you are building yours?