@eerhardt 2020-08-07
> Is this because of the linker?

Yes. You have a non-public, unused method. This is exactly what the linker removes when it is run.