@tarekgh 2020-08-07
Thanks @eerhardt. what is the best way to force the linker to not remove them?