@eerhardt 2020-08-07
Using `DynamicDependency` as listed above will work. But honestly, the use of private reflection here is the issue IMO.