@CodeBlanch 2020-08-07
Below are the perf numbers. I'm going to open a PR to get `DynamicDependency` in here.

@noahfalk 
* As we expected, minor improvement for callers doing foreach. Massive improvement for OTel using the struct Enumerator helper. No change to EnumerateActivityTags.
* Do you want me to PR my perf tests into dotnet/performance?

Without perf improvements (BEFORE):
|                       Method | NumberOfActivities |     Mean |   Error |   StdDev |   Median |      Min |      Max |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------- |------------------- |---------:|--------:|---------:|---------:|---------:|---------:|--------:|------:|------:|----------:|
|        EnumerateActivityTags |               5000 | 313.6 us | 9.05 us | 10.43 us | 312.3 us | 300.9 us | 334.0 us | 33.1633 |     - |     - | 273.44 KB |
|  EnumerateActivityTagObjects |               5000 | 385.1 us | 7.35 us |  7.55 us | 383.1 us | 373.1 us | 403.1 us | 33.1230 |     - |     - | 273.44 KB |
| OTelHelperActivityTagObjects |               5000 | 467.0 us | 9.09 us |  9.34 us | 465.0 us | 449.6 us | 488.3 us |       - |     - |     - | 273.49 KB |
|       EnumerateActivityLinks |               5000 | 546.1 us | 7.17 us |  6.36 us | 545.1 us | 539.0 us | 558.2 us | 47.4138 |     - |     - | 390.63 KB |
|      OTelHelperActivityLinks |               5000 | 618.9 us | 4.77 us |  4.23 us | 618.5 us | 613.3 us | 625.2 us | 45.6731 |     - |     - |  390.7 KB |
|      EnumerateActivityEvents |               5000 | 442.1 us | 4.00 us |  3.55 us | 442.3 us | 437.2 us | 448.8 us | 41.6667 |     - |     - | 351.56 KB |
|     OTelHelperActivityEvents |               5000 | 500.5 us | 8.67 us |  8.11 us | 500.4 us | 488.6 us | 515.2 us | 41.6667 |     - |     - | 351.63 KB |

With perf improvements (AFTER):
|                       Method | NumberOfActivities |     Mean |    Error |   StdDev |   Median |      Min |      Max |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------- |------------------- |---------:|---------:|---------:|---------:|---------:|---------:|--------:|------:|------:|----------:|
|        EnumerateActivityTags (Note: No change to this code path) |               5000 | 334.5 us | 16.09 us | 18.53 us | 328.2 us | 314.3 us | 376.7 us | 32.6705 |     - |     - |  280000 B |
|  EnumerateActivityTagObjects |               5000 | 449.7 us |  7.18 us |  6.37 us | 448.9 us | 440.0 us | 461.9 us | 27.7778 |     - |     - |  240000 B |
| OTelHelperActivityTagObjects |               5000 | 233.1 us | 11.44 us | 13.17 us | 227.6 us | 219.5 us | 259.3 us |       - |     - |     - |       1 B |
|       EnumerateActivityLinks |               5000 | 625.3 us | 17.45 us | 19.39 us | 616.3 us | 605.3 us | 664.4 us | 40.8654 |     - |     - |  360000 B |
|      OTelHelperActivityLinks |               5000 | 258.5 us |  4.89 us |  4.80 us | 257.0 us | 252.7 us | 268.6 us |       - |     - |     - |         - |
|      EnumerateActivityEvents |               5000 | 473.1 us | 27.85 us | 32.07 us | 461.3 us | 441.2 us | 539.3 us | 37.5000 |     - |     - |  320000 B |
|     OTelHelperActivityEvents |               5000 | 251.7 us |  4.69 us |  4.16 us | 250.9 us | 247.0 us | 262.6 us |       - |     - |     - |         - |