@CodeBlanch 2020-08-07
@noahfalk Sorry I was talking specifically about the allocations, not speed. I don't know what's up with the differences in mean. I just ran a new set...

Before:
|                       Method | NumberOfActivities |     Mean |    Error |   StdDev |   Median |      Min |      Max |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------- |------------------- |---------:|---------:|---------:|---------:|---------:|---------:|--------:|------:|------:|----------:|
|        EnumerateActivityTags |               5000 | 311.6 μs | 13.76 μs | 15.85 μs | 311.8 μs | 291.2 μs | 344.9 μs | 32.4519 |     - |     - |  280001 B |
|  EnumerateActivityTagObjects |               5000 | 413.6 μs |  8.77 μs |  9.75 μs | 414.5 μs | 398.8 μs | 432.0 μs | 31.9865 |     - |     - |  280003 B |
| OTelHelperActivityTagObjects |               5000 | 491.5 μs | 24.52 μs | 28.23 μs | 485.2 μs | 460.4 μs | 551.7 μs | 31.8352 |     - |     - |  280057 B |
|       EnumerateActivityLinks |               5000 | 595.1 μs | 18.87 μs | 21.73 μs | 597.7 μs | 560.3 μs | 637.7 μs | 46.2963 |     - |     - |  400000 B |
|      OTelHelperActivityLinks |               5000 | 640.8 μs | 19.84 μs | 22.85 μs | 639.6 μs | 611.1 μs | 686.9 μs | 46.1957 |     - |     - |  400080 B |
|      EnumerateActivityEvents |               5000 | 452.7 μs |  7.69 μs |  6.42 μs | 453.0 μs | 440.6 μs | 465.2 μs | 42.2794 |     - |     - |  360000 B |
|     OTelHelperActivityEvents |               5000 | 556.5 μs | 27.76 μs | 29.71 μs | 562.1 μs | 491.9 μs | 601.4 μs | 42.9688 |     - |     - |  360072 B |
|    EnumerateActivityLinkTags |               5000 | 387.8 μs | 12.29 μs | 13.66 μs | 385.4 μs | 368.2 μs | 417.2 μs | 27.9605 |     - |     - |  240000 B |
|   OTelHelperActivityLinkTags |               5000 | 196.9 μs |  4.99 μs |  5.74 μs | 195.6 μs | 188.7 μs | 209.1 μs |       - |     - |     - |         - |

After:
|                       Method | NumberOfActivities |     Mean |    Error |   StdDev |   Median |      Min |      Max |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------- |------------------- |---------:|---------:|---------:|---------:|---------:|---------:|--------:|------:|------:|----------:|
|        EnumerateActivityTags |               5000 | 315.9 μs |  7.30 μs |  8.41 μs | 315.4 μs | 299.6 μs | 332.2 μs | 33.0882 |     - |     - |  280001 B |
|  EnumerateActivityTagObjects |               5000 | 414.6 μs |  4.45 μs |  3.95 μs | 414.8 μs | 406.8 μs | 420.3 μs | 27.0270 |     - |     - |  240000 B |
| OTelHelperActivityTagObjects |               5000 | 222.6 μs | 16.22 μs | 18.68 μs | 223.2 μs | 191.9 μs | 255.6 μs |       - |     - |     - |       1 B |
|       EnumerateActivityLinks |               5000 | 663.1 μs |  9.76 μs |  9.13 μs | 663.7 μs | 647.7 μs | 677.7 μs | 40.7609 |     - |     - |  360000 B |
|      OTelHelperActivityLinks |               5000 | 273.8 μs |  5.85 μs |  6.51 μs | 273.1 μs | 264.3 μs | 289.4 μs |       - |     - |     - |         - |
|      EnumerateActivityEvents |               5000 | 461.4 μs |  9.92 μs | 11.02 μs | 462.0 μs | 443.7 μs | 484.4 μs | 37.1094 |     - |     - |  320000 B |
|     OTelHelperActivityEvents |               5000 | 254.4 μs |  5.46 μs |  6.28 μs | 254.6 μs | 245.8 μs | 270.3 μs |       - |     - |     - |         - |
|    EnumerateActivityLinkTags |               5000 | 385.2 μs |  5.47 μs |  4.57 μs | 386.4 μs | 373.4 μs | 390.0 μs | 27.4390 |     - |     - |  240000 B |
|   OTelHelperActivityLinkTags |               5000 | 195.5 μs |  2.28 μs |  1.90 μs | 195.7 μs | 191.9 μs | 199.4 μs |       - |     - |     - |         - |

Some of them are more or less identical (311.6 μs vs 315.9 μs), some are a bit off. I don't have an explanation for this currently. Could just be what my system is doing while the benchmarks run, could be a real issue? The change itself is essentially returning an explicit enumerator vs letting the compiler make one, any idea why there would be a measurable difference in that?

I'm not an expert in statistics, but if the variance falls within the error, they should be considered ~equal? 🤷